# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: ClickClack API
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
paths:
...
  /api/workspaces/{workspace_id}:
    get:
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                type: object
                required: [workspace]
                properties:
                  workspace:
                    $ref: "#/components/schemas/Workspace"
    patch:
      operationId: updateWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
      responses:
        "200":
          description: Updated workspace profile
          content:
            application/json:
              schema:
                type: object
                required: [workspace]
                properties:
                  workspace:
                    $ref: "#/components/schemas/Workspace"
                  event:
                    type: object
        "403":
          description: Workspace manager permission required
    delete:
      operationId: deleteWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "204":
          description: Workspace metadata deleted; queued upload-object cleanup may continue asynchronously
        "403":
          description: Workspace owner permission required
...
  /api/workspaces/{workspace_id}/app-installations:
    get:
      operationId: listAppInstallations
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "200":
          description: Active app installations for a workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppInstallationListResponse"
    post:
      operationId: createAppInstallation
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppInstallationRequest"
      responses:
        "201":
          description: Installed app binding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppInstallationResponse"
        "403":
          description: Workspace manager permission required
...
  /api/uploads/{upload_id}:
    get:
      operationId: getUpload
      parameters:
        - name: upload_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upload bytes
...
  /api/dms/{conversation_id}:
    get:
      operationId: getDirectConversation
      parameters:
        - $ref: "#/components/parameters/conversation_id"
      responses:
        "200":
          description: Direct conversation
    delete:
      operationId: hideDirectConversation
      parameters:
        - $ref: "#/components/parameters/conversation_id"
      responses:
        "200":
          description: Closed direct conversation for the current user
...
  /api/bots/self/gallery-actions:
    put:
      operationId: setGalleryActions
      parameters: []
      responses:
        '200':
          description: Current durable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryActionRegistration'
        '202':
          description: Durably queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryActionRegistration'
        '400':
          description: Invalid bounded request
        '403':
          description: Actor, source, destination, installation, or scope denied
        '409':
          description: Identity payload or revision conflict
        '410':
          description: Expired, revoked, or unavailable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GalleryActionRegistration'
  /api/workspaces/{workspace_id}/gallery-actions:
    get:
      operationId: listGalleryActions
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
      - name: source_upload_id
        in: query
        required: true
        schema:
          type: string
      - name: destination_id
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Current durable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryActionDiscoveryResponse'
        '202':
          description: Durably queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryActionDiscoveryResponse'
        '400':
          description: Invalid bounded request
        '403':
          description: Actor, source, destination, installation, or scope denied
        '409':
          description: Identity payload or revision conflict
        '410':
          description: Expired, revoked, or unavailable
  /api/workspaces/{workspace_id}/gallery-actions/open:
    post:
      operationId: openGalleryAction
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Current durable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryActionResult'
        '202':
          description: Durably queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryActionResult'
        '400':
          description: Invalid bounded request
        '403':
          description: Actor, source, destination, installation, or scope denied
        '409':
          description: Identity payload or revision conflict
        '410':
          description: Expired, revoked, or unavailable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GalleryActionOpen'
  /api/gallery-actions/sessions/{session_id}:
    get:
      operationId: statusGalleryAction
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      - name: request_id
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Current durable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryActionResult'
        '202':
          description: Durably queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryActionResult'
        '400':
          description: Invalid bounded request
        '403':
          description: Actor, source, destination, installation, or scope denied
        '409':
          description: Identity payload or revision conflict
        '410':
          description: Expired, revoked, or unavailable
...
  /api/bots/self/gallery-actions/requests/{request_id}/response:
    post:
      operationId: respondGalleryAction
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Current durable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryActionReplyResult'
        '202':
          description: Durably queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryActionReplyResult'
        '400':
          description: Invalid bounded request
        '403':
          description: Actor, source, destination, installation, or scope denied
        '409':
          description: Identity payload or revision conflict
        '410':
          description: Expired, revoked, or unavailable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GalleryActionReply'
components:
  parameters:
    workspace_id:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
...
    conversation_id:
      name: conversation_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    GalleryActionRegistration:
      type: object
      additionalProperties: false
      required:
      - installation_id
      - gallery_actions
      properties:
        installation_id:
          type: string
          pattern: ^[a-zA-Z0-9_.-]{1,128}$
        gallery_actions:
          type: array
          items:
            $ref: '#/components/schemas/GalleryActionDescriptor'
          maxItems: 32
...
    GalleryActionDiscoveryResponse:
      type: object
      additionalProperties: false
      required:
      - gallery_actions
      properties:
        gallery_actions:
          type: array
          items:
            $ref: '#/components/schemas/GalleryActionDiscovery'
...
    GalleryActionResult:
      type: object
      additionalProperties: false
      required:
      - session
      - request
      properties:
        session:
          $ref: '#/components/schemas/GalleryActionSession'
        request:
          $ref: '#/components/schemas/GalleryActionRequest'
    GalleryActionReplyResult:
      type: object
      additionalProperties: false
      required:
      - request
      properties:
        request:
          $ref: '#/components/schemas/GalleryActionRequest'
...
    CreateAppInstallationRequest:
      type: object
      required: [app_slug, bot_user_id]
      properties:
        app_slug:
          type: string
        display_name:
          type: string
          maxLength: 80
        bot_user_id:
          type: string
        config:
          type: object
          additionalProperties: true
        setup_nonce:
          type: string
          minLength: 16
          maxLength: 128
          description: Retry key that returns the original installation instead of creating a duplicate.
...
    AppInstallationResponse:
      type: object
      required: [app_installation]
      properties:
        app_installation:
          $ref: "#/components/schemas/AppInstallation"
    AppInstallationListResponse:
      type: object
      required: [app_installations]
      properties:
        app_installations:
          type: array
          items:
            $ref: "#/components/schemas/AppInstallation"
...
    UpdateWorkspaceRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        slug:
          type: string
        icon_url:
          type: string
...
    Workspace:
      type: object
      required: [id, route_id, name, slug, icon_url, created_at]
      properties:
        id:
          type: string
        route_id:
          type: string
          description: Immutable public route ID used in app URLs.
        name:
          type: string
        slug:
          type: string
        icon_url:
          type: string
        created_at:
          type: string
          format: date-time
        role:
          type: string
          enum: [owner, moderator, member, guest, bot]
...